requirements.txt
django>=5.0,<5.2
gunicorn
python-decouple
psycopg[binary,pool]
dj-database-url
requests
whitenoise
django-allauth[socialaccount]
django-allauth-ui
django-widget-tweaks
slippers
stripe